Clean up components folder, move SideNav to its own folder
