+++ src/Room/Room.tsx
@@ -3,7 +3,7 @@ import { useMutation } from '@apollo/react-hooks'
 import { useParams } from 'react-router-dom'
 import { Flex } from 'rebass'
 
-import { SideNav } from 'components'
+import SideNav from 'SideNav'
 import { useWebsocketContext } from 'Context'
 
 import VolumeContextProvider from 'Player/VolumeContextProvider'
