+++ src/SideNav/index.ts
@@ -1,0 +1,3 @@
+import { SideNav } from './SideNav'
+
+export default SideNav
