+++ src/components/Container.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Box } from 'rebass'
 
-const Container: React.FC = ({ children }) => (
+export const Container: React.FC = ({ children }) => (
   <Box
     sx={{
       maxWidth: '1200px',
@@ -12,5 +12,3 @@ const Container: React.FC = ({ children }) => (
     {children}
   </Box>
 )
-
-export default Container
