+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export * from './SideNav'
+export { Container } from './Container'
